+++ internal/client/service.go
@@ -265,6 +265,7 @@ type Service interface {
 	CreateArtifact(ctx context.Context, req *CreateArtifactRequest) (*Artifact, error)
 	PatchArtifact(ctx context.Context, id string, req *PatchArtifactRequest) (*Artifact, error)
 	GetArtifact(ctx context.Context, id string) (*Artifact, error)
+	ListArtifacts(ctx context.Context, req *ListArtifactsRequest) ([]Artifact, error)
 	DeleteArtifactRepository(ctx context.Context, id string) error
 
 	// Workload (Workload API)